// src/app/shared/interfaces/language.interface.ts
export interface Language {
    code: string; // Código del idioma (por ejemplo, 'en', 'es', 'fr')
    name: string; // Nombre completo del idioma (por ejemplo, 'English', 'Español', 'Français')
    flag: string; // URL de la bandera del idioma (por ejemplo, 'https://www.countryflags.io/en/flat/64.png')
}